control-service: async deployment deletion
Why:
As part of VEP-2272, we need to introduce a process for synchronizing data jobs from the database to Kubernetes.
Currently, the process lacks the ability to delete deployments asynchronously.

What:
We have introduced an asynchronous method for deployment deletion. It is based on the logic that
if the desired deployment is null and the actual deployment is not null, the process performs deployment deletion.

Testing done:
Integration tests.

+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/datajobs/DataJobsDeploymentController.java
@@ -32,6 +32,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
+import static com.vmware.taurus.service.deploy.DataJobDeploymentPropertiesConfig.*;
+
 /**
  * REST controller for operations on data job deployments
  *
@@ -65,7 +67,15 @@ public class DataJobsDeploymentController implements DataJobsDeploymentApi {
     if (jobsService.jobWithTeamExists(jobName, teamName)) {
       // TODO: deploymentId not implemented
       if (jobName != null) {
-        deploymentService.deleteDeployment(jobName);
+        if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.K8S)) {
+          deploymentService.deleteDeployment(jobName);
+
+        }
+
+        if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.DB)) {
+          deploymentServiceV2.deleteDesiredDeployment(jobName);
+        }
+
         return ResponseEntity.accepted().build();
       }
       return ResponseEntity.notFound().build();
@@ -84,11 +94,16 @@ public class DataJobsDeploymentController implements DataJobsDeploymentApi {
       if (job.isPresent()) {
         var jobDeployment =
             ToModelApiConverter.toJobDeployment(teamName, jobName, dataJobDeployment);
-        deploymentService.patchDeployment(job.get(), jobDeployment);
+
+        if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.K8S)) {
+            deploymentService.patchDeployment(job.get(), jobDeployment);
+        }
+
         if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.DB)) {
           deploymentServiceV2.patchDesiredDbDeployment(
               job.get(), jobDeployment, operationContext.getUser());
         }
+
         return ResponseEntity.accepted().build();
       }
     }
@@ -101,8 +116,12 @@ public class DataJobsDeploymentController implements DataJobsDeploymentApi {
     if (jobsService.jobWithTeamExists(jobName, teamName)) {
       // TODO: deploymentId and mode not implemented
       List<DataJobDeploymentStatus> deployments = Collections.emptyList();
-      Optional<JobDeploymentStatus> jobDeploymentStatus =
-          deploymentService.readDeployment(jobName.toLowerCase());
+      Optional<JobDeploymentStatus> jobDeploymentStatus = Optional.empty();
+      if (dataJobDeploymentPropertiesConfig
+              .getReadDataSource()
+              .equals(ReadFrom.K8S)) {
+        jobDeploymentStatus = deploymentService.readDeployment(jobName.toLowerCase());
+      }
       if (jobDeploymentStatus.isPresent()) {
         deployments =
             Arrays.asList(ToApiModelConverter.toDataJobDeploymentStatus(jobDeploymentStatus.get()));
@@ -117,8 +136,12 @@ public class DataJobsDeploymentController implements DataJobsDeploymentApi {
       String teamName, String jobName, String deploymentId) {
     if (jobsService.jobWithTeamExists(jobName, teamName)) {
       // TODO: deploymentId are not implemented.
-      Optional<JobDeploymentStatus> jobDeploymentStatus =
-          deploymentService.readDeployment(jobName.toLowerCase());
+      Optional<JobDeploymentStatus> jobDeploymentStatus = Optional.empty();
+      if (dataJobDeploymentPropertiesConfig
+              .getReadDataSource()
+              .equals(ReadFrom.K8S)) {
+        jobDeploymentStatus = deploymentService.readDeployment(jobName.toLowerCase());
+      }
       if (jobDeploymentStatus.isPresent()) {
         return ResponseEntity.ok(
             ToApiModelConverter.toDataJobDeploymentStatus(jobDeploymentStatus.get()));
@@ -141,17 +164,24 @@ public class DataJobsDeploymentController implements DataJobsDeploymentApi {
       if (job.isPresent()) {
         var jobDeployment =
             ToModelApiConverter.toJobDeployment(teamName, jobName.toLowerCase(), dataJobDeployment);
-        // TODO: Consider using a Task-oriented API approach
-        deploymentService.updateDeployment(
-            job.get(),
-            jobDeployment,
-            sendNotification,
-            operationContext.getUser(),
-            operationContext.getOpId());
-        if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.DB)) {
-          deploymentServiceV2.updateDesiredDbDeployment(
-              job.get(), jobDeployment, operationContext.getUser());
+
+        if (dataJobDeploymentPropertiesConfig
+                .getWriteTos()
+                .contains(WriteTo.K8S)) {
+          // TODO: Consider using a Task-oriented API approach
+          deploymentService.updateDeployment(
+                  job.get(),
+                  jobDeployment,
+                  sendNotification,
+                  operationContext.getUser(),
+                  operationContext.getOpId());
         }
+
+        if (dataJobDeploymentPropertiesConfig.getWriteTos().contains(WriteTo.DB)) {
+            deploymentServiceV2.updateDesiredDbDeployment(
+                    job.get(), jobDeployment, operationContext.getUser());
+        }
+
         return ResponseEntity.accepted().build();
       }
     }
